pyproject.toml
[project]
name = "gems-exchange-mcp-server"
version = "1.0.0"
description = "MCP server for GEMS Exchange agricultural and environmental data APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["endpoints", "server.py", "__init__.py", "__main__.py"]